+++ CHANGELOG.md
@@ -1,3 +1,62 @@
+## [0.1.21](https://github.com/uptrace/opentelemetry-go-extra/compare/v0.1.20...v0.1.21) (2023-02-09)
+
+
+
+## [0.1.20](https://github.com/uptrace/opentelemetry-go-extra/compare/v0.1.19...v0.1.20) (2023-02-01)
+
+
+
+## [0.1.19](https://github.com/uptrace/opentelemetry-go-extra/compare/v0.1.18...v0.1.19) (2023-01-31)
+
+
+
+## [0.1.18](https://github.com/uptrace/opentelemetry-go-extra/compare/v0.1.17...v0.1.18) (2023-01-11)
+
+
+### Bug Fixes
+
+* **otelgorm:** restore original context after "after" callback ([7cd8508](https://github.com/uptrace/opentelemetry-go-extra/commit/7cd850876ad1dfaf4ee7efd5b3d226d75060d7c9))
+* otelsql attr passthrough when reporting db stats ([6172588](https://github.com/uptrace/opentelemetry-go-extra/commit/617258842465e7f6bf085f5f8c53bd0cd864f474))
+
+
+### Features
+
+* **otelgorm:** ignore dry run spans by default ([323b5a9](https://github.com/uptrace/opentelemetry-go-extra/commit/323b5a99e338125fc77216a274bfd09cb932fb3a))
+* **otelgorm:** optionally ignore dryrun callbacks ([5239b0d](https://github.com/uptrace/opentelemetry-go-extra/commit/5239b0d90875322fa57a07338a8ac47e74be50e2))
+
+
+
+## [0.1.17](https://github.com/uptrace/opentelemetry-go-extra/compare/v0.1.16...v0.1.17) (2022-10-07)
+
+
+
+## [0.1.16](https://github.com/uptrace/opentelemetry-go-extra/compare/v0.1.15...v0.1.16) (2022-09-05)
+
+
+### Bug Fixes
+
+* change SugaredLoggerWithCtx.Desuger to return LoggerWithCtx ([12f061e](https://github.com/uptrace/opentelemetry-go-extra/commit/12f061e380d1e2a383207614b139b03e51e946bf))
+* copy skipCaller in With. Fixes [#62](https://github.com/uptrace/opentelemetry-go-extra/issues/62) ([4b58eac](https://github.com/uptrace/opentelemetry-go-extra/commit/4b58eac802c1cd9f564db06a7f832f7e417455c7))
+* **otelzap:** do not panic on invalid input ([#69](https://github.com/uptrace/opentelemetry-go-extra/issues/69)) ([a41206e](https://github.com/uptrace/opentelemetry-go-extra/commit/a41206ed2c49fc7c817f470f61c5082b124cf267))
+
+
+### Features
+
+* **otelzap:** add extra fields option ([aa0e1a0](https://github.com/uptrace/opentelemetry-go-extra/commit/aa0e1a011e1dd8340143e95d1ab2e99300e7dc0b))
+* **otelzap:** make the caller depth configurable ([38a6a03](https://github.com/uptrace/opentelemetry-go-extra/commit/38a6a03114bcaf15e53c5965eaf03b9be8ff0631))
+
+
+
+## [0.1.15](https://github.com/uptrace/opentelemetry-go-extra/compare/v0.1.14...v0.1.15) (2022-07-25)
+
+
+### Bug Fixes
+
+* **otelzap:** add missing DebugwContext function ([ea2b33a](https://github.com/uptrace/opentelemetry-go-extra/commit/ea2b33aa1e368a61b65fa31f4cf31ac794cd06ba))
+* **otelzap:** fix WithTraceIDField for Debugw-like API ([c937b65](https://github.com/uptrace/opentelemetry-go-extra/commit/c937b653c20110be254e2232a4fc22d3de700c94))
+
+
+
 ## [0.1.14](https://github.com/uptrace/opentelemetry-go-extra/compare/v0.1.13...v0.1.14) (2022-05-28)